Initial version with basic functionality
Adds a `MarkStash` command that toggles a floating window with your
marks, let's you navigate through them with j/k and then navigates to
the mark when you hit enter.

It's super basic but I think it'll be effective as a starting point for
figuring out what I want to do with this.