Auto navigate to mark if you hit the key
This is a bit of an ergonomic improvement, and it has some quirk to it,
but basically if you have a mark list like:
```
Window Marks
------------
'  :2
a  end:51

Global Marks
------------
B  lua/markstash/ui.lua:14
C  lua/markstash/ui.lua:65
````

And you hit the `B` key, you'll jump to ui.lua:14, or if you hit `a`,
you'll jump to end:51

The situation where this doesn't work is if you have a mark that
conflicts with a bind like j/k, then for now we just ignore it.

Not sure how that should be handled.